Export, Delete関数を切り出し
